Fix GCC sign-compare: cast -1 to AgentType and int comparison for enemy_type
- main_headless.c: cast -1 to (AgentType) in ternary
- evaluator.c: cast -1 to (AgentType) in assignment
- runner.c: cast enemy_type to (int) for >= 0 comparison

+++ src/sim/runner.c
@@ -49,7 +49,7 @@ void runner_run(const RunConfig* rc, Metrics* metrics) {
 
     Agent enemy;
     Agent* enemy_ptr = NULL;
-    if (rc->enemy_type >= 0) {
+    if ((int)rc->enemy_type >= 0) {
         agent_init(&enemy, rc->enemy_type);
         enemy_ptr = &enemy;
     }
